rename swap_in to swap_in_segwit
