+++ libs/sdk-core/src/error.rs
@@ -6,7 +6,7 @@ use thiserror::Error;
 
 use crate::{
     bitcoin::util::bip32, node_api::NodeError, persist::error::PersistError,
-    swap_in::error::SwapError, swap_out::error::ReverseSwapError,
+    swap_in_segwit::error::SwapError, swap_out::error::ReverseSwapError,
 };
 
 pub type SdkResult<T, E = SdkError> = Result<T, E>;
