+++ libs/sdk-core/src/swap_in_segwit/error.rs
